Update read/written registers for x86 loop instructions (#2799)
* loop

This fixes a bug in 64-bit mode where eip and ecx are read/written
instead of rip and rcx. However, it introduces the bug that using the
size prefix override (0x67e210) still reads/writes rip and rcx instead
of eip and ecx. There is currently no way to solve this issue because
'insn->detail->x86.prefixes' isn't populated at the time when the
register translation is happening in 'X86_get_insn_id'.

* loope

When the address size prefix override is present, 'rcx' is incorrectly
reported as being read/written instead of 'ecx'.

* loopne

As with loope when the address size prefix override is present, 'rcx'
is incorrectly reported as being read/written instead of 'ecx'.